08-01-2020 01:14 PM

+++ Rojmat_Boot/src/main/java/com/rajesh/model/User.java
@@ -31,7 +31,7 @@ public class User extends BaseEntity{
 	@Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "user_id")
-    private int user_id;
+    private Long user_id;
 	
 	@Column(name="shopname")
 	@Length(min = 3, message = "*Your shopname must have at least 3 characters")
@@ -68,10 +68,10 @@ public class User extends BaseEntity{
 	@JoinTable(name = "user_role", joinColumns = @JoinColumn(name = "user_id"), inverseJoinColumns = @JoinColumn(name = "role_id"))
 	private Set<Role> roles;
 
-	public int getId() {
+	public Long getId() {
 		return user_id;
 	}
-	public void setId(int user_id) {
+	public void setId(Long user_id) {
 		this.user_id = user_id;
 	}
 	public String getShopname() {
